Made Luna Lens obligatory
- Luna Lens is now obligatory instead of being a safety strat.

The reasons for this are :
- It works for purifications (piledrivers & ectoplasm)
- It can activate levers
- It can be used for White Forest's Bomber Frame mission
- We can do Nidhoggr's 1st phase as it hits cannons and enemies
- It also works for the Dark Castle skip with Hoop Frame

This means that for all these we do not consume any energy by using the Luna Lens, which in turn reduces how often we need to replenish energy along the run.

Although the biggest impact should be Nidhoggr's 1st phase, as being able to use the solar gun without having to worry about energy is makes managing summons & cannons easier.