Changed when getting Einherjar bike part
- Changed when getting the Einherjar bike part in Ancient Tree.

We were waiting for the chest to spawn after the Earth Armor trap to get it. That was slow.

We now go directly through the gate as soon as it opens.

Since we have to go through this room again on the way back with the coffin : we will open the chest while walking on it with the coffin.