New IGT & Solar Bank interests exploit
- Updated the IGT & Solar Bank interests exploit in San Miguel 1.
- Changed the date for the RNG manip in White Forest.
- Removed previously planned Clock Tower IGT change in San Miguel 3.
- Updated screenshots.

In San Miguel 1 we now use the Clock Tower & a soft-reset to change IGT twice. This should always give at least 3900 Solls (planned amount for the new Any% route).

This is the best moment to do this because we still haven't spent any Solls to buy items : a bigger amount of Solls is affected by the Solar Bank interests which results in getting more Solls from changing IGT.

The IGT change during the RNG manip in White Forest will give much less Solls, however in cases where a small amount of extra Solls is needed due to low interests rates during San Miguel 1, it should be enough to allow the run to continue.