Added 2nd Rapier in San Miguel 1
- Added buying a 2nd Rapier in San Miguel 1 for 200 Solls.

With the new routes in Any% and Any% Sensor-less, it's not necessary to go back to San Miguel before Pirate Island because we'll buy the bananas for it before White Forest.

Beige previously went back to San Miguel before Pirate Island just to forge Epee because Rapier was being used during White Forest and Django didn't meet the level requirement to equip Epee during it.

This trip to San Miguel before Pirate Island is now skipped because with the 2nd Rapier we can forge Epee in San Miguel before White Forest.
This will save time at the cost of needing a bit more Solls at the start of the game.