Added Interest rate manip on New Game
- Added the new manip for getting 18% interest rate during the prologue with Sabata.
- Changed IGT set for the RNG manip in White Forest due to needing to change IGT one less time.
- Updated imgs.

This new manip saves at least 1 minute by itself due to only needed to change IGT once in San Miguel 1.

On top of this, this gives ~7900 Solls which allows for more items to be bought (Power Nuts for Pirate Island, Bearnuts for Lost Paradise).

This interest rate manip is also very easy due to being done at the start of a New Game and only having to hold B during the prologue as Sabata.